#39 [rename] test util 파일 이름 변
